Extract Auction class from main script
